runtime: Reduce the scope of CheckJNI
CheckJNI wasn't anonymous so that it could friend JniEnvExt. Add
accessors for the CheckJNI parts of JniEnvExt and make CheckJNI and
related classes anonymous. This is equivalent to making functions
within the scope of a file static and preferred now by cppstyle.
Make the force_copy_ field that could be const, const.
Test: m -j32; booted emulator.

Change-Id: I7e740d3b9bf4a3fedd30c4f44a852af67898d2f0
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>