Ensure that art_sigsegv_fault is never inlined
We were incorrectly inlining the art_sigsegv_fault function into
FaultManager::HandleFault. This is a problem because we want native
debuggers to break on this function instead of stopping on SIGSEGV,
since we use those for various internal functions. By setting the
art_sigsegv_fault function to be NO_INLINE we should not have this
problem.

Test: Manual inspection of libart.so
Bug: 77528455
Change-Id: I77753cf79966011d7bfbea056bb4efc3f55d64df
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>